[refactor] GameControllerからBall関連とPlayer関連を責務分割
