+++ Assets/Scripts/BallManager.cs
@@ -1,0 +1,98 @@
+using System;
+using UnityEngine;
+
+public class BallManager : MonoBehaviour
+{
+    public event Action<Ball> OnSpawnBall;
+
+    [SerializeField] private GameObject _ballPrefab;
+    [SerializeField] private Vector2 _ballInitialOffset;
+
+    private const float BALL_RESPAWN_SPEED = 0.05f;     // ボールを再生成する下限速度
+    private const float BALL_RESPAWN_TIMEOUT = 3.0f;    // ボールが下限速度を何秒間下回ったら再生成するか
+
+    private Ball _currentBall = null;
+    private float _respawnTimer = 0;
+    private bool _isSelfTurn = true;
+    private bool _isKickedOff = false;
+
+    public void SpawnBall()
+    {
+        if (_currentBall != null)
+        {
+            Destroy(_currentBall.gameObject);
+        }
+
+        float offsetX = _isSelfTurn ? _ballInitialOffset.x : -_ballInitialOffset.x;
+        var ballPosition = new Vector3(offsetX, _ballInitialOffset.y, 0);
+        GameObject ballObject = Instantiate(_ballPrefab, ballPosition, Quaternion.identity);
+
+        _currentBall = ballObject.GetComponent<Ball>();
+        _currentBall.OnTouch += OnTouchBall;
+        _isKickedOff = false;
+
+        OnSpawnBall?.Invoke(_currentBall);
+    }
+
+    public void InactivateCurrentBall()
+    {
+        if (_currentBall == null)
+        {
+            throw new InvalidOperationException("CurrentBallがnullのため、非表示にできません。");
+        }
+
+        _currentBall.Inactivate();
+    }
+
+    public void SetTurnPlayer(bool isSelf)
+    {
+        _isSelfTurn = isSelf;
+    }
+
+    void Update()
+    {
+        HandleBallRespawn();
+    }
+
+    private void HandleBallRespawn()
+    {
+        if (_currentBall == null || !_isKickedOff)
+        {
+            return;
+        }
+
+        if (_currentBall.GetCurrentSpeed() > BALL_RESPAWN_SPEED)
+        {
+            ResetRespawnTimer();
+            return;
+        }
+
+        _respawnTimer += Time.deltaTime;
+
+        if (_respawnTimer > BALL_RESPAWN_TIMEOUT)
+        {
+            SpawnBall();
+        }
+    }
+
+    private void OnTouchBall(Collision collision)
+    {
+        _isKickedOff = true;
+
+        if (collision.gameObject.CompareTag("Rod"))
+        {
+            SoundManager.Instance.PlaySE("se_kick_ball");
+        }
+        else if (collision.gameObject.CompareTag("Wall"))
+        {
+            SoundManager.Instance.PlaySE("se_collision");
+        }
+
+        ResetRespawnTimer();
+    }
+
+    private void ResetRespawnTimer()
+    {
+        _respawnTimer = 0;
+    }
+}
